Changed all of the economy stuff to doubles and not longs. Also added the notification for Minebay. Minebay is on the brink of completion.

+++ main/java/com/minelife/economy/BillHandler.java
@@ -16,7 +16,7 @@ public abstract class BillHandler {
 
     public abstract void update();
 
-    public abstract void pay(Billing.Bill bill, long amount);
+    public abstract void pay(Billing.Bill bill, double amount);
 
     public abstract void delete();
