+++ main/java/com/minelife/economy/packet/PacketDeposit.java
@@ -12,23 +12,23 @@ import net.minecraft.entity.player.EntityPlayerMP;
 
 public class PacketDeposit implements IMessage {
 
-    private long amount;
+    private double amount;
 
     public PacketDeposit() {
     }
 
-    public PacketDeposit(long amount) {
+    public PacketDeposit(double amount) {
         this.amount = amount;
     }
 
     @Override
     public void fromBytes(ByteBuf buf) {
-        this.amount = buf.readLong();
+        this.amount = buf.readDouble();
     }
 
     @Override
     public void toBytes(ByteBuf buf) {
-        buf.writeLong(this.amount);
+        buf.writeDouble(this.amount);
     }
 
     public static class Handler implements IMessageHandler<PacketDeposit, IMessage> {
